==> CODE_notebooks/1.-recipes_dataset-PREPROCESSING.ipynb <==
# RECIPES REVIEW 

# RECOMMENDER SYSTEM WITH MATRIX FACTORIZATION & ML<br>

The following project implements a collaborative recommender system with matrix factorization and machine learning methods for food recipe recommendations. The code takes two datasets, preprocess them, and calculates the weights of a factorized matrix using ML  algorithm.<br>
- **NOTE:** Project inspired by Google Recommender Systems Online Course : https://developers.google.com/machine-learning/recommendation

# Index
- [ 1 - Preprocessing](#1)
- [ 2 - Model Training](#2)
- [ 3 - Predictions](#3)
- [ 4 - Recommendation](#4)

# 1.- PREPROCESSING

## Libraries

##########################################################################################################################

import numpy       as     np
import pandas      as     pd
import tensorflow  as     tf
from   tensorflow  import keras

## Import Datasets
- RAW_recipes
- RAW_interactions

##########################################################################################################################

recipes_id = pd.read_csv        ('raw_data/RAW_recipes.csv', usecols=['id','name'])
recipes_id = recipes_id         [['id','name']]
recipes_id = recipes_id.rename  (columns={'id':'recipe_id', 'name':'name_of_recipe'})

recipes_id

##########################################################################################################################

reviews = pd.read_csv(('raw_data/RAW_interactions.csv'), usecols=['user_id','recipe_id','rating'])

reviews

## Filter Datasets<br>
The users/recipes with little/negligible interactions are removed to create a denser less sparse Utility Matrix:<br>
- Users with 7 reviews or less
- Recipes with 7 reviews or less

##########################################################################################################################

# Calculate the number of recipes rated by each user

user_NumRecipes_count = reviews.groupby('user_id')['recipe_id'].count()         
user_NumRecipes_count

# Creates a list with 'users' that have rated > 7 recipes

users_filtered = user_NumRecipes_count[user_NumRecipes_count > 7].index

# Maintain in 'reviews - user-id' the 'user_id's' contained in 'users_filtered'

reviews_filtered = reviews[reviews['user_id'].isin(users_filtered)]

##########################################################################################################################

# Calculate the number of times each recipe was rated

recipe_NumUsers_count = reviews_filtered.groupby('recipe_id')['user_id'].count() 
recipe_NumUsers_count

# Creates a list with 'recipe_ids' that have been rated > 7 times

recipes_filtered = recipe_NumUsers_count[recipe_NumUsers_count > 7].index

# Maintain in 'reviews_filtered - recipe_id' the 'recipe_id's' contained in 'recipe_NumUsers_count'

reviews_filtered = reviews_filtered[reviews_filtered['recipe_id'].isin(recipes_filtered)]

##########################################################################################################################

reviews_filtered['rating'].value_counts()

# Remove rows with rating value equal to 0

reviews_filtered = reviews_filtered[reviews_filtered['rating'] != 0]
reviews_filtered['rating'].value_counts()

## Utility Matrix

##########################################################################################################################

UtilityMatrix = reviews_filtered.pivot(index='recipe_id', columns= 'user_id', values='rating')
UtilityMatrix = UtilityMatrix.fillna(0)

##########################################################################################################################

# From utility matrix, Get the unique values in the index and put them in a list

recipes_UtilityMatrix = UtilityMatrix.index.unique().tolist()
recipes_UtilityMatrix = pd.DataFrame(recipes_UtilityMatrix, columns=['recipe_id'])
recipes_UtilityMatrix

#########################################################################################################################

# Concatenate DataFrames based on 'key_column' while including values from both DataFrames

recipes_UtilityMatrix =  pd.merge(   recipes_UtilityMatrix,   recipes_id,    on='recipe_id',   how='inner'  )
recipes_UtilityMatrix

###########################################################################################################################
###########################################################################################################################
###########################################################################################################################
###########################################################################################################################
###########################################################################################################################
###########################################################################################################################
###########################################################################################################################

# HERE

###########################################################################################################################
###########################################################################################################################
###########################################################################################################################
###########################################################################################################################
###########################################################################################################################
###########################################################################################################################
###########################################################################################################################

num_recipes        =   recipes_UtilityMatrix.shape[0]
my_recipes_review  =   np.zeros( num_recipes )

## INTRODUCE YOUR RECIPE REVIEWS  HERE<br>
1.- Check the data frame titled **recipes_UtilityMatrix** <br>
2.- **Find** the name of the recipe you want to review; and **log** its corresponding **index number**<br>

    
**Example:**<br>

1. **Recipe Name:** _"almond fudge banana cake"_<br> 
2. **Index ID:** "3"

**In order to add a review of 5 to this recipe, introduce:**<br>

**my_recipes_review[3]      =**    5 
    
    


recipes_UtilityMatrix

## Add your Review HERE<br>
In a scale from **1 to 5**
- **1=** Least you enoyed 
- **5=** Most  you enjoyed

recipes_UtilityMatrix.iloc[3]


#  Introduce Recipe Index       Review              Recipe Name

my_recipes_review[3]            =  5          #  almond fudge banana cake


my_recipes_review[7]            =  3          #  chicken pot pie lasagna
my_recipes_review[12]           =  2          #  chicken fried brown rice                                    
my_recipes_review[5207]         =  4          #  sesame sunflower poppy flax seed buttermilk bread                          
my_recipes_review[7777]         =  3          #  eloise s easy sugar cookies         
my_recipes_review[9093]         =  5          #  mexi cashews                                        
my_recipes_review[11457]        =  2          #  homemade flour tortillas                         
my_recipes_review[12123]        =  5          #  gingerbread truffles                   
my_recipes_review[15003]        =  1          #  vegan soft sugar cookies               
my_recipes_review[17782]        =  3          #  light italian feta omelet                            
my_recipes_review[20001]        =  4          #  bananas and milk                                                            
my_recipes_review[20808]        =  3          #  ragu ratatouille hummus plate  ragu                                      

print('\nMy Reviews:\n')
for i in range(len(my_recipes_review)):
    if my_recipes_review[i] > 0 :
        print(f'I Reviewed:  {my_recipes_review[i]}      to Recipe:  {recipes_UtilityMatrix.loc[i,"name_of_recipe"]}');

#  It will create a new list my_index_reviewed that contains 
#  the indices of the recipes in my_recipes_review 
#  that have a rating number

my_index_reviewed = [i for i in range(len(my_recipes_review)) if my_recipes_review[i] > 0] 

## Update Utility Matrix<br>
YOUR Rersonal Reviews are added to the Utility Matrix

######################################################################################################

UtilityMatrix_Reviewed = UtilityMatrix.astype(bool)
UtilityMatrix_Reviewed = UtilityMatrix_Reviewed.astype(int)

UtilityMatrix          =  UtilityMatrix.to_numpy()
UtilityMatrix_Reviewed =  UtilityMatrix_Reviewed.to_numpy()

######################################################################################################

# Add new Reviews to the existing utility matrix

UtilityMatrix          = np.c_[ my_recipes_review, UtilityMatrix ]
UtilityMatrix_Reviewed = np.c_[ (my_recipes_review != 0).astype(int), UtilityMatrix_Reviewed ]

## Normalize Utility Matrix<br>
Mean is subrtracted from actual Utility Matrix to normalize values

######################################################################################################

# Calculate the average rating review for each recipe row
# Normalize the Dataset

mean_recipe              =  (np.sum(UtilityMatrix*UtilityMatrix_Reviewed,axis=1)/(np.sum(UtilityMatrix_Reviewed, axis=1)+1e-12)).reshape(-1,1)
normalized_UtilityMatrix =  UtilityMatrix - np.multiply(mean_recipe, UtilityMatrix_Reviewed)    

## Save Files<br> 

######################################################################################################

my_recipes_review = pd.DataFrame(my_recipes_review)
my_recipes_review.to_csv('my_recipes_review.csv',index= False)

recipes_UtilityMatrix = pd.DataFrame(recipes_UtilityMatrix)
recipes_UtilityMatrix.to_csv('recipes_UtilityMatrix.csv',index= False)

my_index_reviewed = pd.DataFrame(my_index_reviewed)
my_index_reviewed.to_csv('my_index_reviewed.csv',index= False)

mean_recipe = pd.DataFrame(mean_recipe)
mean_recipe.to_csv('mean_recipe.csv',index= False)
###########################################################################################################################

# It may take some time

UtilityMatrix_Reviewed = np.transpose(UtilityMatrix_Reviewed)
UtilityMatrix_Reviewed = pd.DataFrame(UtilityMatrix_Reviewed)
UtilityMatrix_Reviewed.to_csv('UtilityMatrix_Reviewed.csv',index= False, chunksize=1000)

# It may take some time

UtilityMatrix = np.transpose(UtilityMatrix)
UtilityMatrix = pd.DataFrame(UtilityMatrix)
UtilityMatrix.to_csv('UtilityMatrix.csv',index= False, chunksize=1000)

# It may take some time

normalized_UtilityMatrix  = np.transpose(normalized_UtilityMatrix )
normalized_UtilityMatrix  = pd.DataFrame(normalized_UtilityMatrix )
normalized_UtilityMatrix .to_csv('normalized_UtilityMatrix .csv',index= False, chunksize=1000)

###########################################################################################################################

###########################################################################################################################

###########################################################################################################################

# Continue with TRAINING notebook
