==> CODE_notebooks/3.-recIpes_dataset-PREDICTIONS.ipynb <==
# 3.- PREDICTIONS

## Libraries

##########################################################################################################################

import sys
import numpy
import numpy        as     np
import pandas       as     pd
import tensorflow   as     tf
from   tensorflow   import keras

numpy.set_printoptions(threshold=20)

## Import Information

##########################################################################################################################

chunk_size     =  4162                             # 5 chunks of 4162 rows  =  20810 recipes

X              =  pd.read_csv  ( 'X.csv', chunksize = chunk_size )
W_myReview     =  pd.read_csv  ( 'W_myReview.csv'                ) 
b_myReview     =  pd.read_csv  ( 'b_myReview.csv'                ) 
mean_recipe    =  pd.read_csv  ( 'mean_recipe.csv'               )

b_myReview  =  b_myReview. to_numpy()
W_myReview  =  W_myReview. to_numpy()
mean_recipe =  mean_recipe.to_numpy().reshape(-1,1)

## Predictions Calculation

##########################################################################################################################


num_recipes    =  chunk_size
num_users      =  W_myReview.shape[1]              # 1 user
num_features   =  W_myReview.shape[0]              # 100 features

Final_Predic   =  np.empty((0,num_users), dtype=np.float64)

num_of_chunks  =  5   # 5 chunks of 4162 = 20810 out of 20810 .......one round
chunk_count    =  1

for CHUNK_1 in zip(X):
    
    print(f"Training chunk {chunk_count}")
    
    CHUNK_1 = pd.DataFrame(CHUNK_1[0])
    X_chunk_values = CHUNK_1.values
      
    if chunk_count == num_of_chunks:
        
        # Handle the last chunk with a different shape

        last_chunk_size = CHUNK_1.shape[0]
        predictions = np.empty((last_chunk_size, num_users), dtype=np.float64)
    
    else:
        
        predictions = np.empty((num_recipes, num_users), dtype=np.float64)
     
        
    predictions = np.matmul(X_chunk_values, W_myReview) + b_myReview
    
    
    
    Final_Predic = np.concatenate((Final_Predic, predictions), axis=0)

    print(Final_Predic.shape)
                             
                             
                             
        
    if chunk_count == num_of_chunks:
        break
    
    chunk_count += 1                           

##########################################################################################################################

#restore the mean

Final_Predic   =  Final_Predic + mean_recipe

my_predictions =  Final_Predic[:,0]

## Save Files

##########################################################################################################################

my_predictions =  pd.DataFrame(my_predictions)
my_predictions.to_csv('my_predictions.csv',index= False)

##########################################################################################################################

##########################################################################################################################

##########################################################################################################################

# Continue with RECOMMENDATIONS notebook
