==> CODE_notebooks/4.-recIpes_dataset-RECOMMENDATIONS.ipynb <==
# 4.- RECOMMENDATIONS

## Libraries

##########################################################################################################################

import sys
import numpy
import numpy        as     np
import pandas       as     pd
import tensorflow   as     tf
from   tensorflow   import keras

numpy.set_printoptions(threshold=20)

## Import Information

##########################################################################################################################

my_predictions          =  pd.read_csv( 'my_predictions.csv'        )
my_recipes_review       =  pd.read_csv( 'my_recipes_review.csv'     )
my_index_reviewed       =  pd.read_csv( 'my_index_reviewed.csv'     )
recipes_UtilityMatrix   =  pd.read_csv( 'recipes_UtilityMatrix.csv' )

my_predictions               =  my_predictions.to_numpy()
my_predictions               =  np.squeeze(my_predictions)

my_recipes_review            =  my_recipes_review.to_numpy()
my_index_reviewed            =  my_index_reviewed['0'].tolist()
recipes_UtilityMatrix_list   =  recipes_UtilityMatrix['name_of_recipe'].tolist()

## Predictions

##########################################################################################################################

# sort predictions

ix  =   tf.argsort(my_predictions, direction='DESCENDING')

##########################################################################################################################

print('\n\nMy Original Reviews          Predicted Reviews             Recipe Name\n')
for i in range(len(my_recipes_review)):
    if my_recipes_review[i] > 0:
        print(f'  My Review: {my_recipes_review[i]}             Predicted: {my_predictions[i]:0.2f}           {recipes_UtilityMatrix_list[i]}')

## Top 12 Recommendations

##########################################################################################################################

print('\nBased on Your Personal Reviews\n')
print('\nTop 12 Recommended Recipes:\n')
for i in range(12):
    j = ix[i]
    if j not in my_index_reviewed:
        print(f'Predicted Review:   {my_predictions[j]:0.2f}       Recipe:  {recipes_UtilityMatrix_list[j]}')

## Different PREDICTIONS

### Do you want to know how much will you enjoy a RECIPE you've  never tried before?<br> 
    1. introduce the Index of the Recipe below


recipes_UtilityMatrix

### I want to know how much I may enjoy "biscuits and gravy" recipe

Recipe_Index=   1212

recipes_UtilityMatrix.iloc[Recipe_Index]


print(f'\n\nRecipe Name:            {recipes_UtilityMatrix_list[Recipe_Index]} \nEnojoyment Level:       {my_predictions[Recipe_Index]:0.2f}  \n  ')

##########################################################################################################################

##########################################################################################################################

##########################################################################################################################

# END OF PROGRAM
